# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Encurtador de Links",
    "description": "API Criada para a disciplina Desenvolvimento Back-End para Aplicações Mobile da Pós Unyleya ",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [{
    "name": "shortener",
    "description": "Operações relacionadas aos links"
    
  }],
  "schemes" : ["http"],
  "consumes" : ["application/json"],
  "produces" : ["application/json"],
  
    "paths": {
      "/": {
        "get": {
          "tags": [
            "shortener"
          ],
          "summary": "Retorna a página inicial",
          "description": "Retorna a página inicial",
          "operationId": "getHome",
          "responses": {
            "200": {
              "description": "Sucesso",
              "schema": {
                "type": "string"
              }
            },
            "400": {
              "description": "Erro",
              "schema": {
                "type": "string"
              }
            }
          }
        }
      
      },
      "/short/{id}": {
        "get": {
          "tags": [
            "shortener"
          ],
          "summary": "Retorna o link encurtado",
          "description": "Retorna o link encurtado",
          "operationId": "getShort",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID do link encurtado",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Sucesso",
              "schema": {
                "message": "string",
                "shortUrl": "string",
                "seedUrl": "string"
              }
            },
            "400": {
              "description": "Erro",
              "schema": {
                "type": "string"
              }
            }
          }
        }
      },
      "/short": {
        "post": {
          "tags": [
            "shortener"
          ],
          "summary": "Encurta o link",
          "description": "Encurta o link",
          "operationId": "postShort",
          "consumes": ["application/json"],
          "parameters": [
            {
              "name": "url",
              "in": "body",
              "description": "Link a ser encurtado",
              "required": true,
              "type": "object"
            }
          ],
          "responses": {
            "200": {
              "description": "Sucesso",
              "schema": {
                 "message": "string",
                 "urlId": "string",
                 "shortUrl": "string"
              },
              "properties": {
                "url" : "string"
              }
            },
            "201": {
              "description":"URL já encurtada",
              "schema":{
                "message": "string",
                "urlId": "string"
              }
            },
            "404": {
              "description":"Not found",
              "schema":{
                "message": "string"
              }
            },
            "400": {
              "description": "Erro",
              "schema": {
                "type": "string"
              }
            }
          }
        }
      }
  }
    
  
}